Moved mongo-related files to eva-lib
