+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/repository/VariantEntityRepository.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.server.repository;
+package uk.ac.ebi.eva.lib.repository;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
